spider/vnf_examples/object-detector/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
Flask==2.0.2
flatbuffers==1.12
gast==0.3.3
google-auth
google-auth-oauthlib
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageai==2.1.6
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.2
kubernetes==19.15.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.3.2
numpy==1.19.3
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
Pillow==7.0.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy
six==1.15.0
tensorflow==2.2.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
Flask-Classful==0.14.2
kubernetes==19.15.0
